+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.7 (unreleased)
 
 - Added `--debug` option
+- Added experimental `--format` option
 - Fixed error when Redis has no keys
 
 ## 0.1.6 (2022-09-16)